Return requested options and reply NoBinding to RELEASE
Honor the requested options list.  The client and server does not
currently support all possible options, but at least honor what
the client requests from the server and only return those options.

When the client sends a DH6_RELEASE message, release the IA_NA and
return a REPLY message with the NOBINDING status and the success
code set.