Workaround problem in /usr/include/asm/gcc_intrin.h on ia64
There is a problem with /usr/include/asm/gcc_intrin.h on ia64 on
at least RHEL 5.1, which is my build platform for the moment.  So
to get around that, if I see ia64, I include <linux/in6.h> and then
define struct in6_ifreq (copied from ipv6.h).  Yeah, it's a hack,
but only temporary.