Fix SIGSEGV while parsing saved configuration independently of flex version
Depends on the flex version, dhcp6c segfaults if it tries to read
the saved configuration. The cause is the change of yy_init behavior.
We should not set yy_init ourselves as yy_init value is set to
do the init by default.
The idea for this fix is by Karsten Keil <kkeil@suse.de>

Signed-off-by: Mitsuru Chinen <chinen@wile.yamato.ibm.com>