Remove /etc/radvd.conf file modification code.
The DHCPv6 client should not be updating the radvd(8) conf
file.  Not sure why this functionality was still around, but
it's bad form.  If radvd(8) is also going to run on the same
system, it should be getting the latest information about the
network information via netlink or some other method.