Use the last 4 bytes of the MAC addr to generate IAID.
According to RFC 3315:
Each IA has an IAID, which is chosen to be unique among all
IAIDs for IAs belonging to that client.

However, a machine with multiple nics and similar MAC addrs
chooses duplicate IAIDs.

Currently, the first four bytes of the MAC address are used
to generate the IAID.  In hosts with multiple interfaces, the
MAC address for each interface can vary solely in the last
byte.  The current algorithm guarantees that duplicate IAIDs
will be generated in such situations, which violates the RFC
as quoted above.  Using the last 4 bytes of the MAC address
to generate the IAID should be unique enough for practical
purposes.  Neither of these methods guarantee uniqueness, but
the current algorithm is more prone to collisions than the
suggested one.

Patch from Rob Myers <rob.myers AT gtri DOT gatech DOT edu>