Make dhcp6c work on s390 virtual network interfaces.
We tried to use the dhcp6c on a s390 iucv0 interface,
this will give a segfault with the current code because
we sre still missing some tests.

Program received signal SIGSEGV, Segmentation fault.
0x000002aaaaac1628 in get_duid () from /usr/sbin/dhcp6c

The problem for the segfault is, that it do not check
the return value of the
    len = calculate_duid_len(ifname, &hwtype);
call.

For supporting the hardware itself, more changes are
needed, the the interfaces are from type 256
(ARPHRD_SLIP) so maybe it could work if it does the
same as with ARPHRD_PPP, I will do some testing, if I
get access to the HW.

Patch from kkeil.