Update internal status correctly on RENEW/REBIND recv
When receiving RENEW or REBIND message, current code doesn't
update the internal status of addresses but validate the
addresses.  The reason is addr_flag is set to a value to
validate the addresses although the value is set to update
the address once.  This patch updates the internal status
correctly.

Patch from Mitsuru Chinen <mitch@linux.vnet.ibm.com>