Write server PID file after config file is parsed
If a management app is waiting for the server pid file to
show up, it could falsely think the server is running if
the conf file was bad, moving the writing of the pid file
to just before the main loop elevates this problems.  Move
the PID file writing to after config file parsing, just
before server6_mainloop() is called.

Patch from jkl6648