Delay sending Confirm until the completion DAD for link-local addr.
Confirm message is sent out while DAD for the link-local
address is not complete.  This sometimes causes drop Reply
message from a server as the link-local address stays
tentative state.  We should delay sending a Confirm message
until the DAD for a link-local address is complete.

Patch from Mitsuru Chinen.