Clear request address list when a message is sent out successfully
At client6_recvreply(), request_list variable, which stores request
address list, is initialized in the middle.
dhcp6c retransmits a message when it fails to send the message. In that
case, request address list should be in retransmitted the message.
The variable should be initialized when a message is sent out sucessfully.

Signed-off-by: Mitsuru Chinen <mitch@linux.vnet.ibm.com>
Signed-off-by: David Cantrell <dcantrell@redhat.com>