Need to call init_lease_hashes()
Although dhcp6c uses hash_anchors array, the memory for the array
is not allocated and the array is not initialized. Then, dhcp6c
causes SIGSEGV when it accesses lease_hash_table, which is an alias
of hash_anchors[HT_IPV6LEASE]. Not only dhcp6s but also dhcp6c
must call init_lease_hashes().

Signed-off-by: Mitsuru Chinen <mitch@linux.vnet.ibm.com>
Signed-off-by: David Cantrell <dcantrell@redhat.com>