Only send DNS server list or domain list if the client requests it.
The logic in dhcp6s for filling in the requested options isn't very
robust, so this patch is more like a temporary fix.  Basically, what
happens now is dhcp6s checks to see if the option is requested and
then copies information in, but then a function in common.c is called
to build the reply message and it ignores the option request list
and goes based on whether or not the server is configured with any
of that information.