Start filling out run_script() and begin glib conversion.
This is a mess of a commit, apologies to those who track the source.
I've been working on dhcpv6 a lot lately and the old coding style and
reinvention of a number of basic components was starting to irritate me,
so I've slowly started to move the code over to using glib for basic
things.  This will happen over time, but it means that you'll need glib
in order to use dhcpv6 from now on.

As for run_script(), it's starting to get some structure to it as I
dissect the client code.