Replace dhcpv6_dprintf() with GLib log handling.
Added log_handler() to common.c to handle logging messages.  We can use
the macros and/or the g_log() function to log messages.  This gets rid
of the ugly dhcpv6_dprintf() function.

Also take care of a long standing request to use the date and time
format described here:
    http://www.w3.org/TR/NOTE-datetime

Note, that the ISO 8601 date format is only valid for logging messages
to the console.  When using syslog(3), the system logger handles time
stamps for us.  The ISO 8601 format used for foreground logging in
dhcpv6 is:  YYYY-MM-DDThh:mmTZD (eg 1997-07-16T19:20+01:00)