Convert struct dns_list to use GSList.
Replaced struct dns_list with dns_info_t.  It has two members, both
GSLists.  The first is called 'servers' and is a list of struct
in6_addr elements, which are DNS servers.  The other is called
'domains' and is a list of gchar strings of domain names.

Updated the code to use dns_info_t for local and received DNS info,
which greatly simplifies the code.  The list of server addresses was
a TAILQ structure while the list of domains was just a plain linked
list implemented internally.  Now both lists are GSList types and
a little more readable.