Replace TRACE() macro with g_log() macros.
The TRACE macro was sort of sloppy.  It was defined as fprintf.  The
first argument to the macro was the file descriptor to log too, which
was stderr if you didn't specify a log (dump) file.  Lame.

I've changed all of the TRACE() messages to use g_error() or g_debug(),
which uses the logging structure for the rest of the program.

I also cleaned up the messages a bit, making them look less weird (just
use __func__ to get the function name).