Removed _find_event_by_xid() and _find_event_by_state()
These were used in one place each by a g_slist_find_custom() call.  In
order to enhance readability, replaced these functions with a local loop
and a simple value test.  This is actually more appropriate as we want
to just return NULL if not found, otherwise a pointer to the event.  The
GCompareFunc callbacks should return non-zero values of the value
difference, which makes it suitable for use as a sorting function.  We
don't really need that here, just a simple lookup instead.