Convert remaining structs to typedef types.
For custom types, you may have noticed my conversion to typedef in all
of the header files.  I prefer using typedef for custom types and naming
the type ending with "_t".  I prefer the way the code looks when I do
this, so I wanted them all to be typedefs.