Make str.c functions return fixed strings.
Change the else blocks to log a message about the unknown code or type
and then return a fixed string.  Previously, only if the type or code
was unknown did you have to free what you got back from the string
conversion function.  If it was known, you didn't have to free it.