Add dhcp6_option2envvar() and related code.
For script mode in dhcp6c, we need the ability to convert all options in
the DHCPv6 message to environment variables.  Added
dhcp6_option2envvar() to do that, but realized it would be really useful
to have a list of options that were actually in the reply rather than
relying on structure members containing a value and assuming the option
was set.

Modified dhcp6_get_options() to build up opt_list, which is a new member
of the dhcp6_optinfo_t type.  This is just a GSList of the option codes.
Iterate that in run_script and call dhcp6_option2envvar() for each
option.  Depending on the option, there will be a different way to
gather the option data and turn it in to a string, that's what
dhcp6_option2envvar() is for.

Also renamed DH6OPT_IADDR to DH6OPT_IAADDR because it's always bugged me
because it's inconsistent with the abbreviated names from RFC 3315.

Stubbed out the code for converting options.