Include current time in foreground logging.
If the software is run in the foreground, log messages go to stderr and
the timestamp is given according to this format:

    http://www.w3.org/TR/NOTE-datetime

The current time was always zero, but the date was right.  Changed to
using time(2) and localtime(3) to get the current time stamp and
strftime(3) to format it.  The glib variant doesn't seem to handle
times, only dates.