Modify copy_option() to return the length of the copied option.
Replacing the COPY_OPTION macro with copy_option() left the option
length counter set to zero for the entire run of dhcp6_set_options().
Have copy_option() return the length it just copied in to the buffer,
and keep a running total in dhcp6_set_options().  If we get -1 from
copy_option(), it means failure.